requirements.txt
Django==4.1.1
djangorestframework==3.14.0
requests==2.28.1
gunicorn==20.1.0
psycopg2==2.9.3
django-heroku-postgresify==0.4.1